+++ packages/xchain-ethereum/__tests__/client.test.ts
@@ -1,4 +1,5 @@
-import { ethers, Wallet, providers } from 'ethers'
+import nock from 'nock'
+import { ethers, Wallet, providers, utils } from 'ethers'
 import { formatEther, parseEther } from '@ethersproject/units'
 import { TransactionResponse, TransactionReceipt } from '@ethersproject/abstract-provider'
 import { baseAmount, AssetETH, BaseAmount } from '@xchainjs/xchain-util'
@@ -6,7 +7,7 @@ import { Balances } from '@xchainjs/xchain-client'
 import Client from '../src/client'
 import { ETH_DECIMAL } from '../src/utils'
 import { mockDashboardAddress, mockGetTx } from '../__mocks__/blockchair-api'
-import { mock_eth_gasPrice } from '../__mocks__/etherscan-api'
+import { mock_all_api } from '../__mocks__'
 
 /**
  * Test Data
@@ -18,6 +19,8 @@ const address = '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e'
 const vault = '0x8c2A90D36Ec9F745C9B28B588Cba5e2A978A1656'
 const blockchairUrl = 'https://api.blockchair.com/ethereum/testnet'
 const etherscanUrl = 'https://api-goerli.etherscan.io'
+const goerliInfraUrl = 'https://goerli.infura.io/v3'
+const goerliAlchemyUrl = 'https://eth-goerli.alchemyapi.io/v2'
 const wallet = {
   signingKey: {
     curve: 'secp256k1',
@@ -56,6 +59,14 @@ const txResponse = {
  * Wallet Tests
  */
 describe('Wallets', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('should throw error on bad phrase', () => {
     expect(() => {
       new Client({ phrase: 'bad bad phrase' })
@@ -88,6 +99,14 @@ describe('Wallets', () => {
  * Connectivity Tests (networks/providers)
  */
 describe('Connecting', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('should connect to specified network', async () => {
     const ethClient = new Client({
       network: 'mainnet',
@@ -119,6 +138,14 @@ describe('Connecting', () => {
  * Utils
  */
 describe('Utils', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('should get address', () => {
     const ethClient = new Client({
       network: 'testnet',
@@ -145,6 +172,14 @@ describe('Utils', () => {
 })
 
 describe('Balances', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('gets a balance without address args', async () => {
     const ethClient = new Client({
       network: 'testnet',
@@ -238,6 +273,14 @@ describe('Balances', () => {
 })
 
 describe('Transactions', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('get transaction history', async () => {
     const ethClient = new Client({
       network: 'testnet',
@@ -338,6 +381,62 @@ describe('Transactions', () => {
       },
     })
 
+    mockGetTx(blockchairUrl, '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2', {
+      '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2': {
+        transaction: {
+          block_id: 3888697,
+          id: 3888697000001,
+          index: 1,
+          hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+          date: '2020-12-08',
+          time: '2020-12-08 04:49:56',
+          failed: false,
+          type: 'call',
+          sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+          recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+          call_count: 1,
+          value: '1000',
+          value_usd: 0,
+          internal_value: '0',
+          internal_value_usd: null,
+          fee: '432160000000000',
+          fee_usd: null,
+          gas_used: 21608,
+          gas_limit: 32412,
+          gas_price: 20000000000,
+          input_hex:
+            'a9059cbb0000000000000000000000001d5516bc994a130fc4fa2a6275f5b4ee022b18720000000000000000000000000000000000000000000000000000000004cace80',
+          nonce: 98,
+          v: '2e',
+          r: '8797fe82968c48ccaf7db43785a2a97c6868b187a009f930e8406a052aefdc7b',
+          s: '54dc9740ebf75835bd3aca3de018454251213e13b44fa93ecbf5f17faf16d995',
+        },
+        calls: [
+          {
+            block_id: 3888697,
+            transaction_id: 3888697000001,
+            transaction_hash: '0x0816e0f18643b8a53b52091321954733bc173542f01424f4ca86cbf1d2e567b2',
+            index: '0',
+            depth: 0,
+            date: '2020-12-08',
+            time: '2020-12-08 04:49:56',
+            failed: false,
+            fail_reason: null,
+            type: 'call',
+            sender: '0x4a89644d5dffb825a42a3496e24510424ca01516',
+            recipient: '0xdac17f958d2ee523a2206206994597c13d831ec7',
+            child_call_count: 0,
+            value: '1000',
+            value_usd: 0,
+            transferred: true,
+            input_hex:
+              'a9059cbb0000000000000000000000001d5516bc994a130fc4fa2a6275f5b4ee022b18720000000000000000000000000000000000000000000000000000000004cace80',
+            output_hex: '',
+          },
+        ],
+      },
+    })
+
     const txHistory = await ethClient.getTransactions({
       address: '0xdac17f958d2ee523a2206206994597c13d831ec7',
       limit: 1,
@@ -354,7 +453,7 @@ describe('Transactions', () => {
 
   it('get transaction data', async () => {
     const ethClient = new Client({
-      network: 'testnet',
+      network: 'mainnet',
       phrase,
       blockchairUrl,
     })
@@ -427,53 +526,59 @@ describe('Transactions', () => {
     expect(txData.type).toEqual('transfer')
   })
 
-  it.only('sends a normalTx', async () => {
+  it('sends a normalTx', async () => {
     const ethClient = new Client({
       network: 'testnet',
       phrase,
       blockchairUrl,
     })
 
-    // const mockTx = jest.spyOn(ethClient.getProvider(), 'getBlockNumber')
-    // mockTx.mockImplementation(
-    //   async (): Promise<number> => {
-    //     return Promise.resolve(1)
-    //   },
-    // )
-
-    mock_eth_gasPrice(etherscanUrl, '0xb2d05e00')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_blockNumber', '0x3c6de5')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_getTransactionCount', '0x10')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_gasPrice', '0xb2d05e00')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_estimateGas', '0x5208')
+    mock_all_api(
+      etherscanUrl,
+      goerliInfraUrl,
+      goerliAlchemyUrl,
+      'eth_sendRawTransaction',
+      '0xc2b98a50107fbb6ed9369c552e332f50810eaef04a70c0253b617106f02a674d',
+    )
 
     const txResult = await ethClient.transfer({
       recipient: '0x8ced5ad0d8da4ec211c17355ed3dbfec4cf0e5b9',
       amount: baseAmount(100, ETH_DECIMAL),
     })
-    expect(txResult).toEqual(0)
+    expect(txResult).toEqual('0xc2b98a50107fbb6ed9369c552e332f50810eaef04a70c0253b617106f02a674d')
   })
 
   it('sends a normalTx with special parameters', async () => {
-    const ethClient = new Client({ network: 'mainnet', phrase })
+    const ethClient = new Client({
+      network: 'testnet',
+      phrase,
+      blockchairUrl,
+    })
 
-    const mockTx = jest.spyOn(ethClient.getWallet(), 'sendTransaction')
-    mockTx.mockImplementation(
-      async (_): Promise<TransactionResponse> => {
-        return Promise.resolve(txResponse)
-      },
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_blockNumber', '0x3c6de5')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_getTransactionCount', '0x10')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_gasPrice', '0xb2d05e00')
+    mock_all_api(etherscanUrl, goerliInfraUrl, goerliAlchemyUrl, 'eth_estimateGas', '0x5208')
+    mock_all_api(
+      etherscanUrl,
+      goerliInfraUrl,
+      goerliAlchemyUrl,
+      'eth_sendRawTransaction',
+      '0xe82829e2c2d425c83a093ada3ee58a0680e0fa199f59aae79ee922f7ce90c797',
     )
 
-    await ethClient.normalTx({
+    const txResult = await ethClient.normalTx({
       recipient: ethClient.getAddress(),
-      amount: baseAmount(parseEther('1').toString(), 18),
+      amount: baseAmount(100, ETH_DECIMAL),
       overrides: {
-        nonce: 123,
-        data: '0xdeadbeef',
+        data: utils.toUtf8Bytes('memo'),
       },
     })
-    expect(mockTx).toHaveBeenCalledWith({
-      data: '0xdeadbeef',
-      nonce: 123,
-      to: '0xb8c0c226d6fe17e5d9132741836c3ae82a5b6c4e',
-      value: '1',
-    })
+    expect(txResult.data).toEqual('0x6d656d6f')
   })
 
   it('checks vault and vaultTx', async () => {
@@ -487,6 +592,14 @@ describe('Transactions', () => {
 })
 
 describe('ERC20', () => {
+  beforeEach(() => {
+    nock.disableNetConnect()
+  })
+
+  afterEach(() => {
+    nock.cleanAll()
+  })
+
   it('gets erc 20 balance for a contract without addr', async () => {
     const ethClient = new Client({ network: 'testnet', phrase })
 
@@ -512,7 +625,7 @@ describe('ERC20', () => {
     mockerc20.mockImplementation(async (_): Promise<BaseAmount> => Promise.resolve(baseAmount(100000, 18)))
 
     const gasEstimate = await ethClient.estimateGasERC20Tx({
-      erc20ContractAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
+      assetAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
       recipient: '0x2fe25ca708fc485cf356b2f27399247d91c6edbd',
       amount: 1,
     })
@@ -527,7 +640,7 @@ describe('ERC20', () => {
     mockerc20.mockImplementation(async (_): Promise<TransactionResponse> => Promise.resolve(txResponse))
 
     const txR = await ethClient.erc20Tx({
-      erc20ContractAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
+      assetAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
       recipient: '0x2fe25ca708fc485cf356b2f27399247d91c6edbd',
       amount: 1,
       overrides: {
@@ -537,7 +650,7 @@ describe('ERC20', () => {
 
     expect(txR).toEqual(txResponse)
     expect(mockerc20).toHaveBeenCalledWith({
-      erc20ContractAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
+      assetAddress: '0xc3dbf84Abb494ce5199D5d4D815b10EC29529ff8',
       recipient: '0x2fe25ca708fc485cf356b2f27399247d91c6edbd',
       amount: 1,
       overrides: {